pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "planwise"
version = "0.1.0"
description = "Agentic local issue tracker"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "filelock>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
planwise = "planwise.cli:cli"
pw = "planwise.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/planwise"]
